% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSummary.R
\name{GMSummary}
\alias{GMSummary}
\title{Geometric mean for model evaluation}
\usage{
GMSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{a data frame with columns obs and pred for the observed and predicted outcomes.}

\item{lev}{a character vector of factors levels for the response.}

\item{model}{a character string for the model name (as taken from the method argument of \code{\link{train}}.}
}
\value{
A vector of geometric mean performance estimates.
}
\description{
Provides functionality to evaluate a model using the gemetric mean as described in Galar et al., 2014.
The function can be used in caret's \code{\link{train}} and CAST'S \code{\link{ffs}} to select variables based on the geometric mean.
}
\examples{
\dontrun{
data(iris)
ctrl <- trainControl(method="cv",summaryFunction = GMSummary)
model <- train(iris[,1:4],iris$Species,trControl=ctrl,metric="GM")
model
}
}
\references{
\itemize{
\item Galar, M., Fernandez, A., Barrenechea, E., Herrera, F. (2014): Empowering difficult classes with a similarity-based aggregation in multi-class classification problems. Information Sciences 264: 135-157.
}
}
\seealso{
\code{\link{train}},code{\link{trainControl}},\code{\link{ffs}},
\code{\link{defaultSummary}}
}
\author{
Hanna Meyer, Jannis Gottwald, Thomas Nauß
}
