% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionStats.R
\name{regressionStats}
\alias{regressionStats}
\title{calculate prediction performance statistics of regression models}
\usage{
regressionStats(prd, obs, adj.rsq = TRUE, method = "pearson")
}
\arguments{
\item{prd}{numeric vector of predicted values}

\item{obs}{numeric vector of observed values}

\item{adj.rsq}{logical, whether to return adjusted r-squared.
Defaults to TRUE}

\item{method}{character. Method to use for correlation. See \code{?cor.test}
for details.}
}
\value{
a data frame.
}
\description{
this function calculates prediction performance statistics
between vectors of predicted and observed values, namely
coefficient of determination (Rsq), root mean squared error (RMSE),
mean error (ME), mean absolute error (MAE).
}
\examples{
## create predictions with high accuracy (identical mean),
## but low precision (sd double that of observations). Hence,
## ME should be close to zero and RMSE close to ten.
pred_vals <- sort(rnorm(1000, 200, 20)) # sorting ensures high Rsq
obs_vals <- sort(rnorm(1000, 200, 10))
result <- regressionStats(pred_vals, obs_vals, adj.rsq = FALSE)
result

}
\author{
Tim Appelhans, Hanna Meyer
}
