% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptm-package.R
\docType{package}
\name{sptm}
\alias{sptm}
\alias{sptm-package}
\title{Supporting functions for spatio-temporal models using machine learning}
\description{
This repository is a (currently random) collections of functions that support machine learning model training and evaluation in environmnetal research
}
\details{
Supporting functions for spatio-temporal models using machine learning
}
\author{
Hanna Meyer, Tim Appelhans, Jannis Gottwald, Thomas Nauss
\cr
\emph{Maintainer:} Hanna Meyer \email{hanna.meyer@uni-muenster.de}
}
